Split yi-stock-modes
