+++ stack.yaml
@@ -8,6 +8,7 @@ packages:
 - yi-keymap-vim
 - yi-mode-haskell
 - yi-mode-javascript
+- yi-stock-modes
 - yi-language
 - yi
 extra-deps:
