+++ yi-core/package.yaml
@@ -92,14 +92,10 @@ library:
         - Yi.Main
         - Yi.MiniBuffer
         - Yi.Misc
-        - Yi.Mode.Abella
-        - Yi.Mode.Buffers
-        - Yi.Mode.Compilation
-        - Yi.Mode.GHCi
-        - Yi.Mode.IReader
+        - Yi.Mode.Common
         - Yi.Mode.Interactive
-        - Yi.Mode.Latex
-        - Yi.Modes
+        - Yi.Mode.IReader
+        - Yi.Mode.Compilation
         - Yi.Monad
         - Yi.Paths
         - Yi.PersistentState
