+++ yi-keymap-emacs/yi-keymap-emacs.cabal
@@ -1,4 +1,4 @@
--- This file has been generated from package.yaml by hpack version 0.14.0.
+-- This file has been generated from package.yaml by hpack version 0.14.1.
 --
 -- see: https://github.com/sol/hpack
 
@@ -32,6 +32,7 @@ library
     text,
     transformers-base,
     yi-core,
+    yi-stock-modes,
     yi-language,
     yi-rope
   ghc-options: -Wall -ferror-spans
