+++ yi-mode-haskell/src/Yi/Mode/GHCi.hs
