+++ yi-mode-javascript/src/Yi/Mode/JavaScript.hs
@@ -38,7 +38,7 @@ import           Yi.Keymap                 (Action (..), YiM, topKeymapA)
 import           Yi.Keymap.Keys            (ctrlCh, important, (?>>), (?>>!))
 import           Yi.Lexer.Alex             (AlexState, CharScanner, Tok, commonLexer, lexScanner)
 import           Yi.Lexer.JavaScript       (HlState, TT, Token, alexScanToken, initState)
-import           Yi.Modes                  (anyExtension)
+import           Yi.Mode.Common            (anyExtension)
 import           Yi.Monad                  (gets)
 import qualified Yi.Rope                   as R (fromString, fromText)
 import           Yi.String                 (showT)
