+++ yi-stock-modes/src/Yi/Mode/Buffers.hs
