+++ yi-stock-modes/yi-stock-modes.cabal
@@ -2,14 +2,14 @@
 --
 -- see: https://github.com/sol/hpack
 
-name:           yi-mode-haskell
+name:           yi-stock-modes
 version:        0.13
+synopsis:       Yi editor javascript mode
 category:       Yi
-synopsis:       Yi editor haskell mode
-license:        GPL-2
-maintainer:     Dmitry Ivanov <ethercrow@gmail.com>
 homepage:       https://github.com/yi-editor/yi#readme
 bug-reports:    https://github.com/yi-editor/yi/issues
+maintainer:     Dmitry Ivanov <ethercrow@gmail.com>
+license:        GPL-2
 build-type:     Simple
 cabal-version:  >= 1.10
 
@@ -20,28 +20,23 @@ source-repository head
 library
   hs-source-dirs:
       src
-  default-language: Haskell2010
-  exposed-modules:
-      Yi.Lexer.Haskell
-      Yi.Lexer.LiterateHaskell
-      Yi.Mode.GHCi
-      Yi.Mode.Haskell
-      Yi.Mode.Haskell.Dollarify
-      Yi.Syntax.Haskell
-      Yi.Syntax.Paren
-      Yi.Syntax.Strokes.Haskell
-  other-modules:
-      Paths_yi_mode_haskell
+  ghc-options: -Wall -ferror-spans
   build-depends:
       base >= 4.8
-    , array
     , binary >= 0.7
     , data-default
+    , filepath
     , microlens-platform
+    , semigroups
     , text
     , yi-core
     , yi-language
     , yi-rope
-  ghc-options: -Wall -ferror-spans
-  include-dirs:
-      src/Yi/Lexer
+  exposed-modules:
+      Yi.Modes
+      Yi.Mode.Buffers
+      Yi.Mode.Abella
+      Yi.Mode.Latex
+  other-modules:
+      Paths_yi_stock_modes
+  default-language: Haskell2010
